added functionality to activate/deactivate nonstaffusers as a staff user from the user management screen
