+++ src/index.css
@@ -102,6 +102,18 @@
   .menu-card-text{
     @apply m-auto text-lg font-medium;
   }
+
+  .discreet-button{
+    @apply relative flex items-center justify-center content-center px-2 h-full font-semibold text-lg font-medium transition-all duration-150 ease-linear;
+  }
+
+  .discreet-button-sky{
+    @apply text-sky-200 border-solid border-2 rounded-lg border-sky-600 hover:text-sky-100 transition-all duration-150 ease-linear;
+  }
+
+  .discreet-button-emerald{
+    @apply text-emerald-200 border-solid border-2 rounded-lg border-emerald-600 hover:text-emerald-50 transition-all duration-150 ease-linear;
+  }
 }
 
 body {
